% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{rotate_x}
\alias{rotate_x}
\title{Rotate wrt x axis
also is_rotation checker}
\usage{
rotate_x(theta)
}
\arguments{
\item{theta}{A float}
}
\value{
a 3x3 matrix
}
\description{
Rotate wrt x axis
also is_rotation checker
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
